Added information about the authorization token
The specification purposely does not specify the authorization token
format and structure as well as methods used to secure the token. That
is part of the provieder-specific implementation so that providers can
choose their preferred formats and methods. However, the specification'
was lacking a recommnendation of what a meaningful token should contain.

Signed-off-by: Rudolf J Streif <rstreif@jaguarlandrover.com>