Simplifying Success and Error Responses for VSS Request and Get
The amount of Vehicle data will be huge so that one of API design
principle is to minimize a size of exchange data between vehicle client
and server.

Request messages from client to the server in both of VSS request and
Get have action, path and requestId. Via communication context client
could look for action and path matched with requestId in response
message from the server.