Removed repetition from the introduction: "RESTful web services are out of scope for the first revision of this specification, but could be considered for addition in a later version."
"For security reasons, clients are not able to connect directly to ECUs or to CAN, MOST or LIN networks. All access is via WebSocket. This allows the server to securely control access to vehicle signals."

Removed repetition from "Security and Privacy Considerations - Use of Encryption"

"To support ‘defence in depth’ and a layered security approach, connections between clients and servers are strongly encrypted. This is to make it more difficult for an attacker that has succeeded in installing malicious code on a vehicle to eavesdrop, hijack security tokens or impersonate valid security principals in order to get and set sensitive vehicle signals."