update numeric data-types
Through the reduction of several data-types expressing numbers the implementation in an transport protocol becomes easier.
As all memory used by data-types has to be defined the so called restriction (seen from the stand point of using a data-value) define the necessary memory to be allocated.
The set of pre-defined restriction are introduced to help existing implementations and descriptions to be used within Gen2.